ReactElite-Project
This YAML file sets up a GitHub workflow to build and deploy a Node.js app to an Azure Web App. It defines the build and deployment steps, sets up the Node.js version, uploads the build artifact, downloads it in the deployment job, and deploys it to the Azure Web App using the Azure Web Apps Deploy action. The workflow is triggered on a push to the main branch and can also be triggered manually.